+++ scripts/shipyard_nodeprep.sh
@@ -382,8 +382,8 @@ docker_pull_image() {
     set -e
 }
 
-echo "Configuration:"
-echo "--------------"
+echo "Configuration [Non-Native Docker]:"
+echo "----------------------------------"
 echo "Batch Shipyard version: $version"
 echo "Blobxfer version: $blobxferversion"
 echo "Offer/Sku: $offer $sku"
@@ -397,6 +397,7 @@ echo "Azure File: $azurefile"
 echo "GlusterFS on compute: $gluster_on_compute"
 echo "HPN-SSH: $hpnssh"
 echo "Block on images: $block"
+echo ""
 
 # check sdb1 mount
 check_for_buggy_ntfs_mount
@@ -836,8 +837,8 @@ fi
 
 # retrieve docker images related to data movement
 docker_pull_image alfpark/blobxfer:$blobxferversion
-docker_pull_image alfpark/batch-shipyard:tfm-$version
-docker_pull_image alfpark/batch-shipyard:rjm-$version
+docker_pull_image alfpark/batch-shipyard:${version}-tfm
+docker_pull_image alfpark/batch-shipyard:${version}-rjm
 
 # login to registry servers (do not specify -e as creds have been decrypted)
 ./registry_login.sh
@@ -922,14 +923,14 @@ p2p=$p2p
 EOF
     chmod 600 $envfile
     # pull image
-    docker_pull_image alfpark/batch-shipyard:cascade-$version
+    docker_pull_image alfpark/batch-shipyard:${version}-cascade
     # launch container
     docker run $detached --net=host --env-file $envfile \
         -v /var/run/docker.sock:/var/run/docker.sock \
         -v $AZ_BATCH_NODE_ROOT_DIR:$AZ_BATCH_NODE_ROOT_DIR \
         -w $AZ_BATCH_TASK_WORKING_DIR \
         -p 6881-6891:6881-6891 -p 6881-6891:6881-6891/udp \
-        alfpark/batch-shipyard:cascade-$version &
+        alfpark/batch-shipyard:${version}-cascade &
     cascadepid=$!
 else
     # add timings